Add tox and test against Python 3.4 (#15)
* Move tests to their own module

It's good practice to keep testing code separate from production code,
and there have been talks of not including testing code with the
released packages. This commit achieves that.

* Add Tox to test against different Python versions

We support several Python versions. Putting tox in place ensures a
developer can easily run the tests against all of them.

* Update Travis configuration

We agreed to also support Python 3.4, so this commit adds it to the list
of Python versions to test against.

I also removed explicit dependencies from the Travis configuration - I
think it's too easy for them to become outdated.

* Document Python versions and how to contribute